pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mcutils"
version = "0.1"
authors = [
    { name = "Drew Gochenaur" }
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=5.0",
    "platformdirs>=4.3",
    "meshcore",
    "folium",
    "grpcio>=1.60",
    "protobuf>=4.25"
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.60",
    "mypy>=1.8",
    "mypy-protobuf>=3.6",
    "types-protobuf"
]

[project.scripts]
mcutils = "mcutils.mcutils:main"
server = "mcutils.server:main"
client = "mcutils.client:main"
tcp-server = "mcutils.tcp_server:main"

[tool.black]
line-length = 160

[tool.pyright]
typeCheckingMode = "strict"